Initial commit. Almost fully working!
res_coupled.py gives the same answer as Matlab implementation, so seems to be working properly.
JFNK.py contains the JFNK solver, which passes the unit test but when replacing fsolve in channel_solver.py does not work.
A bug is present in res_nk_rel.py, which is set up to receive coolant temperature at surface values and interpolate to node values, however, res_coupled.py is already giving interpolated values. This double interpolation is underestimating coolant temperatures (biasing to inlet temperature). It should be fixed after the solver works, so that it can be compared with Matlab first.